Overhaul bash config files
Ditch whatever it was I was using before and use
[bash-it](https://github.com/Bash-it/bash-it) instead. bash-it is nice
and modular, and its trivial for me to write my own stuff independent of
it. I've replaced all the .bashrc.os.d bumpf with a single folder
`.bashrc.d` which holds configuration files, all of which are sourced at
runtime, each file should check whether it should run or not dependent
upon the context (e.g. uni machine or not)